spring-boot: move spring-boot:feign-client to spring-cloud:openfeign
