+++ generators/generate-blueprint/internal/lookup-namespaces.spec.ts
@@ -98,12 +98,12 @@ describe('lookupGeneratorsNamespaces', () => {
   "spring-boot",
   "spring-boot:bootstrap",
   "spring-boot:cucumber",
-  "spring-boot:feign-client",
   "spring-boot:jwt",
   "spring-boot:oauth2",
   "spring-boot:websocket",
   "spring-cache",
   "spring-cloud",
+  "spring-cloud:feign-client",
   "spring-cloud:gateway",
   "spring-cloud:kafka",
   "spring-cloud:pulsar",
