+++ generators/spring-cloud/generators/feign-client/cleanup.ts
