+++ generators/spring-cloud/generators/feign-client/generator.spec.ts
