+++ generators/spring-cloud/generators/feign-client/index.ts
