+++ generators/spring-cloud/generators/feign-client/templates/src/main/java/_package_/client/AuthorizedFeignClient.java.ejs
