+++ generators/spring-cloud/generators/feign-client/templates/src/main/java/_package_/config/FeignConfiguration.java.ejs
